Accept docker_insecure_registries as string
In case of single registry, it can be a string also. So just accept it.

I was trying to get the registry as variable, and in the playbook, I was
checking if the variable is set and if it is, was trying to convert it
to a list using .split(). But for some reason, split was not working if
I use along with conditionals i.e {%if a %} {{ a.split()}}{%endif%}
doesnt work where {{a.split()}} without any conditionals work. So I
ended up having this workaround in this role.