Handle docker and containers
Added ability to do
* Install docker on a base host
* Configure docker to use an insecure registry
* Start controller, analytics and agent containers with appropriate
 parameters - this is only work for single node environment at this
 moment. System will get the container image from configured registry or
 from dockerhub (If not provided a registry)

This code added an inventory from which the host inventory as well as
the arguments/variables will be taken

Added a requirements.yml to handle external role dependencies -
currently it use hkumarmk/ansible-docker role to setup docker. The idea
is to use reusable roles which are available in ansible-galaxy or
github. This will enable us to reduce the pain of handling our own code
for everything.

Have a conditional logic in site.yml to run appropriate code in
different environments -

* We would have the ansible code running inside the container - to
 configure the container.
* And the code running outside the container which is responsible to
* setup base host like setting up docker in them, setup the base host,
 start the containers etc