Added vagrant code to start development environment
Ths path add vagrant code to start development environment. This need
vagrant and virtualbox installed on the machine, once that is done, you
just come to the directory where Vagrantfile exist and run vagrant
commands to manage the systems.

This code will enable one to define development environment in a yaml
file including the machines, type of machines, size and other
configuration.