Detect if the code is running in docker or on base host
Detecting where the code is running - whether is in docker or in base
host, is bit complicated when we run docker with --pid=host,
--privileged which is something we are doing for our controller
container because of some issues we see otherwise.

In this case, default ansible fact (ansible_virtualization_type) doesnt
work here as it just talk about what the base host is. So did a
workaround I see to detect - by checking a file /.dockerenv is there
exists or not - this file is created by docker in docker containers. In
future if docker change that behavior we may have to check on this
further. Otherwise I dont see any solution for this.

This patch add this custom fact from custom_facts.yml which will be
called as first task so that the fact (running_in_docker) is available
to all other code.