Added vrouter module compilation in agent container workflow
Now in case of redhat, ansible will start vrouter compiler container to
compile and install the module to the system, and then start the agent
container.

This is required for redhat and ubuntu xenial OS as current agent
container cannot support compiling modules for latest operating systems
because of gcc caps mismatch between parameters in kernel headers and
caps available in gcc version in the container.