Use tag configure to configure system rather than "provsion"
Provision should be a separate task in which we would be provisionnig
various config objects like registering config objects for variuos nodes
like config, control nodes.

Now we are moving running provision_* python scripts to ansible which
should be running with "provision" tag, so renaming current tasks which
areactually configuring the system to tag "configure".

Other reason is that the term as well ansible tag "configure" is much
more widely used for that purpose.