enable_webui_service variable to decide to setup webui
added variable enable_webui_service to determine whether to setup and
enable webui. If enable_webui_service is false, the role will make sure
webui is disabled on the node and if it is true, role will setup and
start webui service.

Also disable redis in case webui is disabled.