Making config_server_list optional when all
controller containers are enabled with config
component, to make the container input config
file backward compatable.

config_server_list is the new variable required
in controller and analytics containers input conf file
to decide on the list of zookeepers/cassandra/rabbit
servers to be used in the container's service config
files. When all controller containers are enabled with
config component controller_list should be considered
as config_server_list.

Closes: https://github.com/Juniper/contrail-ansible/issues/310