Fix errors
The my_*_ip is not getting passed to cassandra or any dependencies
because that variable is set from set_fact task and all tasks are
executed after running dependencies. This is causing failures when we
pass my_*_ip variables to the dependency roles like cassandra.

To solve this, moved them as variable - also see we refer it in multiple
places, so moved defining them to contrail/common.