Avoid unnecessary "skipped" tasks
When conditionally call tasks/include task files, ansible go through
entire tasks and print them as skipped, which is a nuisance as well as
not efficient. This patch is to avoid this by dynamic include using
variables with fileglob in the include path which will only include them
if they exist.

This will also solve any failures due to any issues on variables
substitution due to those optional tasks/roles.