Derive controller_ip from controller_list
Earlier it was other way around, which was mandating providing both
controller_list and controller_ip to make it work in multi-node
environment - even with single controller with separate compute. This
patch will change it - once somebody provide controller_list,
controller_ip will be derived from it(take first ip), if not provided.