Workarounds and fixes
* Cassandra fail to start/restart (even with multiple restarts), it is
 erroring out with some Nullpointer exception, a service "start" was not
working as systemd consider that the service exit normally and thus it
never start again. Workaround was to wait for some time and then do a
restart
* contrail-api systemd service file that is installed by the package had
the port hardcoded to 9100, which doesnt work with the ansible logic -
ansible has logic to run it on 8082 in single node or in case of
external lb, but in this case all configuration is overridden by
hardcoded parameter in the service file. Workaround was to keep systemd
service file within ansible and copy it over.
* Fix a typo