Fix for agent container not working
The package contrail-vrouter-dkms was missing in ubuntu agent packages.
As a result, the package dependency resolution can be ambigious.
It can either resolve and install DPDK or kernel vrouter.
Since we are not supporting DPDK in agent container, explicitly
provide the kernel package contrail-vrouter-dkms as part of agent
packages to resolve the ambiguity.

Change-Id: Ifbd0f922b136db5b13a620a8545fc8b09969d3eb
Closes-Bug: 1692369