fix clustering issues for rabbitmq
- leader/follower status must be detected not from comparison
of hostname with resolved name from IP. it can be non relevant.
leader detection should be done by comparision leader IP with
all systems' IP-s

- native RabbitMQ cluster uses first nodename from the clustered_nodes
in configuration file. so leader is used in the ansible must be equal
to first element that is stored to configuration

- rabbitmq nodes must be short names
unless RABBITMQ_USE_LONGNAME is not true

- when IP-s are used to provision RabbitMQ is not from default interface
then RabbitMQ couldn't connect to the cluster as used node names for clustering are
different from the running node names. this patchset defines current riabbit node
in the environment and use it also for vhost and user creation.

Closes-Bug: #1697013
Closes-Bug: #1697039
Change-Id: Iace530fad0f6af8193a098b07c7edfb65b8bfebb