Fixed for contrail-kube-manager service issue
Status of contrail-kube-manager.service in ubuntu16.04 would be
checked before it would be restarted. During the boot time,
contrail-kube-manager.service would be in in-active. Eventhough
it is a valid state for systemctl, ansible would take it as failure
and failed to start the kube-manager. Added addtional check to
validate the failure from ansible point of view.

Change-Id: Ic585c4bcd9bd8682da4d8743d9a0a2921d8fa160
Closes-bug: #1719084