Restart kube-manager service forever, on failure.
The systemd defualt is give up on restart of kube-manager, if the
process fails 5 times within 10 secs. But the desired behavior is
to keep retrying start of kubemanager. This commit sets the systemd
restart settings such that kubemanager failure will not hit the
failure threshold of systemd and hence systemd will keep restarting
kube-manager on failure.

Change-Id: Ic1ee7d781739b5a857d0c02cf6929f1dddcb72af
Partial-bug: #1729247